# core/docs/openapi.yaml
openapi: 3.1.0
info:
  title: CauseFlow API
  description: |
    CauseFlow — AI SRE platform. REST API for incident ingestion, triage,
    investigation, remediation, memory/knowledge, integrations, billing,
    tenant/user/team management, notifications, audit, widget/portal, and
    skills.

    Architecture: Modular monolith (Modlito) + Clean Architecture.
    Stack: TypeScript + Hono + ElectroDB (DynamoDB single-table) + Redis + SQS.

    ## Authentication
    Most dashboard endpoints use `bearerAuth` (Clerk JWT).
    Ingestion webhook endpoints accept either a shared webhook secret or a
    tenant API key (`X-API-Key`). Widget endpoints use `X-API-Key`.
    Stripe, Clerk and Composio webhooks verify provider signatures and are
    mounted as public.
  version: 0.1.0
  contact:
    name: CauseFlow
servers:
  - url: https://api.causeflow.io
    description: Production
  - url: https://staging-api.causeflow.io
    description: Staging
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /v1/integrations/credentials:
    post:
      tags: [Integration]
      summary: Connect a credential-based integration (admin/owner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, credentials]
              properties:
                provider: { type: string }
                credentials: { type: object, additionalProperties: true }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Integration' }
  /v1/integrations/credentials/{type}:
    delete:
      tags: [Integration]
      summary: Disconnect a credential-based integration (admin/owner)
      parameters: [ { in: path, name: type, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: Disconnected
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /v1/integrations/test-connection:
    post:
      tags: [Integration]
      summary: Test an AWS AssumeRole (admin/owner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type: { type: string, enum: [cloudwatch, aws] }
                roleArn: { type: string }
      responses:
        '200':
          description: Test result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
  /v1/integrations/connect:
    post:
      tags: [Integration]
      summary: Initiate OAuth connection via Composio (admin/owner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, redirectUrl]
              properties:
                provider: { type: string }
                redirectUrl: { type: string, format: uri }
      responses:
        '200':
          description: OAuth URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  authUrl: { type: string, format: uri }
        '503': { description: Composio not configured }
  /v1/integrations/connect/{provider}:
    delete:
      tags: [Integration]
      summary: Revoke a Composio OAuth connection (admin/owner)
      parameters: [ { in: path, name: provider, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: Revoked
          content:
            application/json:
              schema: { type: object, properties: { success: { type: boolean } } }
        '503': { description: Composio not configured }

  # ─────────────────────────── Triggers (Composio) ───────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk-issued JWT. Tenant context derived from `o.id` claim.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Tenant-scoped API key for programmatic and widget access.
    webhookSecret:
      type: apiKey
      in: header
      name: X-Webhook-Secret
      description: Shared webhook secret for ingestion endpoints.
    webhookSignature:
      type: apiKey
      in: header
      name: webhook-signature
      description: HMAC signature header used by Composio webhooks.
    stripeSignature:
      type: apiKey
      in: header
      name: stripe-signature
      description: Stripe webhook signature header.
    svixSignature:
      type: apiKey
      in: header
      name: svix-signature
      description: Svix signature for Clerk webhooks.
...
  schemas:
...
    Integration:
      type: object
      properties:
        type: { type: string }
        provider: { type: string }
        name: { type: string }
        status: { type: string, enum: [connected, disconnected, error] }
        source: { type: string, enum: [credential, composio] }
        createdAt: { type: string, format: date-time }
...